job-service/requirements.txt
fastapi
uvicorn[standard]
motor
aio-pika
python-jose[cryptography]
pydantic[email]
python-multipart